+++ src/path/axisSpecifiers.ts
@@ -1,5 +1,6 @@
 import * as _ from "lodash";
 
+import { parentOf } from "../internal/treeUtils";
 import { TreeNode } from "../TreeNode";
 import { AxisSpecifier } from "./pathExpression";
 
@@ -46,3 +47,37 @@ export function allDescendants(tn: TreeNode): TreeNode[] {
     return (tn.$children || []).concat(
         _.flatMap(tn.$children.map(kid => allDescendants(kid))));
 }
+
+export const FollowingSiblingAxisSpecifier: AxisSpecifier = {
+
+    type: "following-sibling",
+
+    follow(tn: TreeNode, root: TreeNode): TreeNode[] {
+        const parent = parentOf(tn, root);
+        if (!parent) {
+            return [];
+        }
+        const kids = parent.$children;
+        const index = kids.indexOf(tn);
+        return (index >= 0 && kids.length > index + 1) ?
+            [kids[index + 1]] :
+            [];
+    },
+};
+
+export const PrecedingSiblingAxisSpecifier: AxisSpecifier = {
+
+    type: "preceding-sibling",
+
+    follow(tn: TreeNode, root: TreeNode): TreeNode[] {
+        const parent = parentOf(tn, root);
+        if (!parent) {
+            return [];
+        }
+        const kids = parent.$children;
+        const index = kids.indexOf(tn);
+        return (index > 0) ?
+            [kids[index - 1]] :
+            [];
+    },
+};